fix: Refactored Workout view component
